move BinnedSpectrum to KERNEL
