+++ src/openms/include/OpenMS/COMPARISON/SPECTRA/BinnedSpectrumCompareFunctor.h
@@ -10,7 +10,7 @@
 
 #include <OpenMS/DATASTRUCTURES/DefaultParamHandler.h>
 #include <OpenMS/CONCEPT/Exception.h>
-#include <OpenMS/COMPARISON/SPECTRA/BinnedSpectrum.h>
+#include <OpenMS/KERNEL/BinnedSpectrum.h>
 
 #include <cmath>
