+++ src/tests/class_tests/openms/source/BinnedSharedPeakCount_test.cpp
@@ -11,7 +11,7 @@
 
 ///////////////////////////
 #include <OpenMS/COMPARISON/SPECTRA/BinnedSharedPeakCount.h>
-#include <OpenMS/COMPARISON/SPECTRA/BinnedSpectrum.h>
+#include <OpenMS/KERNEL/BinnedSpectrum.h>
 #include <OpenMS/FORMAT/DTAFile.h>
 ///////////////////////////
