+++ src/tests/class_tests/openms/source/BinnedSpectrum_test.cpp
@@ -10,7 +10,7 @@
 #include <OpenMS/test_config.h>
 
 ///////////////////////////
-#include <OpenMS/COMPARISON/SPECTRA/BinnedSpectrum.h>
+#include <OpenMS/KERNEL/BinnedSpectrum.h>
 ///////////////////////////
 
 #include <Eigen/Sparse>
