+++ src/tests/class_tests/openms/source/ClusterHierarchical_test.cpp
@@ -12,7 +12,7 @@
 ///////////////////////////
 #include <OpenMS/ML/CLUSTERING/ClusterHierarchical.h>
 #include <OpenMS/ML/CLUSTERING/SingleLinkage.h>
-#include <OpenMS/COMPARISON/SPECTRA/BinnedSpectrum.h>
+#include <OpenMS/KERNEL/BinnedSpectrum.h>
 #include <OpenMS/COMPARISON/SPECTRA/BinnedSharedPeakCount.h>
 #include <OpenMS/SYSTEM/File.h>
 #include <OpenMS/FORMAT/DTAFile.h>
